+++ NEWS.md
@@ -6,3 +6,4 @@
 # gellipsoid 0.7.1
 
 * Added a `NEWS.md` file to track changes to the package.
+* Fixed a CRAN nit regarding `options(warn=-1)`